Upgrade settings activity to latest AppCompatActivity and PreferenceFragment
